// src/node_service/data/zod_generated.ts
export const UserSchema = z.object({
  id: z.number().int(),
  username: z.string(),
  name: z.string(),
  isAdmin: z.boolean(),
  grannyId: z.number().int().nullable(),
})
// ...
export const GrandmaSchema = z.object({
  id: z.number().int(),
  username: z.string(),
  name: z.string(),
  description: z.string(),
  pictureUrl: z.string(),
  timeReply: z.number().int(),
  rating: z.number(),
  verified: z.boolean(),
})

export type Grandma = z.infer<typeof GrandmaSchema>

/////////////////////////////////////////
// MEAL SCHEMA
/////////////////////////////////////////

export const MealSchema = z.object({
  id: z.number().int(),
  name: z.string(),
  price: z.number().int(),
  rating: z.number(),
  pictureUrl: z.string(),
  description: z.string(),
  grannyId: z.number().int(),
  cookedBy: z.string(),
  cookedByName: z.string(),
})
// ...
export const OrderStatusSchema = z.object({
  id: z.number().int(),
  name: z.string(),
})